+++ formulas/pool_quoter.go
@@ -398,6 +398,8 @@ func (pm *PoolManager) buildQuoter(pool common.Address, formulaID int) (pq PoolQ
 		if hasTokens {
 			if p := newPlatypusPool(pool, trackedReader, tokens, pm.evmCaller); p != nil { return wrapAndCache(p) }
 		}
+	case FormulaWooFi:
+		if p := newWooFiPool(pool, trackedReader); p != nil { return wrapAndCache(p) }
 	}
 	return nil
 }